fix: 프로젝트 구조 변경
