--- src/main/java/com/example/daangn/domain/PostImage.java
@@ -1,27 +1,0 @@
-package com.example.daangn.domain;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "PostImages")
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class PostImage {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long piuid;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false)
-    private Post post;
-
-    @Column(nullable = false)
-    private String postImg;
-}
