--- src/main/java/com/example/daangn/domain/ProductImage.java
@@ -1,28 +1,0 @@
-package com.example.daangn.domain;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@Entity
-@Table(name = "ProductImages")
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ProductImage {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long piuid;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "puid", nullable = false)
-    private Product product;
-
-    @Column(nullable = false)
-    private String productImg;
-}
