--- src/main/java/com/example/daangn/domain/ProductLike.java
@@ -1,29 +1,0 @@
-package com.example.daangn.domain;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@Entity
-@Table(name = "ProductLikes")
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ProductLike {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long pluid;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id", nullable = false)
-    private Product product;
-}
