+++ src/main/java/com/example/daangn/domain/chat/entity/Chatting.java
@@ -1,0 +1,45 @@
+package com.example.daangn.domain.chat.entity;
+
+import com.example.daangn.domain.product.entity.Product;
+import com.example.daangn.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Table(name = "Chattings")
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Chatting {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long cuid;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "product_id", nullable = false)
+    private Product product;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "sell_user_id", nullable = false)
+    private User sellUser;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "buy_user_id", nullable = false)
+    private User buyUser;
+
+    private LocalDateTime created;
+
+    private LocalDateTime lastest;
+
+    @OneToMany(mappedBy = "chatting", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<ChattingLog> logs = new ArrayList<>();
+}
