+++ src/main/java/com/example/daangn/domain/location/entity/Location.java
@@ -1,0 +1,37 @@
+package com.example.daangn.domain.location.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Entity
+@Table(name = "Locations")
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Location {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long luid;
+
+    @Column(nullable = false)
+    private String si;
+
+    @Column(nullable = false)
+    private String gugun;
+
+    @Column(nullable = false)
+    private String location;
+
+    @Column(nullable = false, precision = 10, scale = 7)
+    private BigDecimal lat;
+
+    @Column(nullable = false, precision = 10, scale = 7)
+    private BigDecimal lng;
+}
