+++ src/main/java/com/example/daangn/domain/post/entity/PostCommentLike.java
@@ -1,0 +1,37 @@
+package com.example.daangn.domain.post.entity;
+
+import com.example.daangn.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "PostCommentLikes")
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PostCommentLike {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long pcluid;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id", nullable = false)
+    private Post post;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "comment_id", nullable = false)
+    private PostComment comment;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "recomment_id")
+    private PostReComment recomment;
+}
