+++ src/main/java/com/example/daangn/domain/product/entity/Product.java
@@ -1,0 +1,63 @@
+package com.example.daangn.domain.product.entity;
+
+import com.example.daangn.domain.chat.entity.Chatting;
+import com.example.daangn.domain.location.entity.Location;
+import com.example.daangn.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Table(name = "Products")
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Product {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long puid;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "location_id", nullable = false)
+    private Location location;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @Column(nullable = false)
+    private String title;
+
+    private String category;
+
+    @Column(nullable = false, columnDefinition = "TEXT")
+    private String content;
+
+    private String dealType;
+
+    private Integer price;
+
+    @Column(nullable = false)
+    private LocalDateTime created;
+
+    private Integer views;
+
+    private Boolean isSell;
+
+    @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<ProductImage> images = new ArrayList<>();
+
+    @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<ProductLike> likes = new ArrayList<>();
+
+    @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Chatting> chattings = new ArrayList<>();
+}
