+++ src/main/java/com/example/daangn/domain/product/entity/ProductLike.java
@@ -1,0 +1,27 @@
+package com.example.daangn.domain.product.entity;
+
+import com.example.daangn.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.*;
+
+
+@Entity
+@Table(name = "ProductLikes")
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProductLike {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long pluid;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "product_id", nullable = false)
+    private Product product;
+}
