--- src/main/java/com/example/daangn/repository/ProductImageRepository.java
@@ -1,14 +1,0 @@
-package com.example.daangn.repository;
-
-import com.example.daangn.domain.Product;
-import com.example.daangn.domain.ProductImage;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
-    List<ProductImage> findByProduct(Product product);
-    void deleteByProduct(Product product);
-}
